unsuccessful decoupling (posted on forum for help)

+++ cpp/cpp/math/ColorMath.h
@@ -29,14 +29,14 @@ public:
         int r = (rgbPixel) >> 16 & 0xFF;
         int g = (rgbPixel) >> 8 & 0xFF;
         int b = rgbPixel & 0xFF;
-        double hue, saturation, brightness;
+        float hue, saturation, brightness;
         
-        double cmax = (r > g) ? r : g;
+        float cmax = (r > g) ? r : g;
         if (b > cmax) cmax = b;
-        double cmin = (r < g) ? r : g;
+        float cmin = (r < g) ? r : g;
         if (b < cmin) cmin = b;
         
-        brightness = cmax / 255.0;
+        brightness = cmax / 255;
         if (cmax != 0)
             saturation = (cmax - cmin) / cmax;
         else
@@ -44,18 +44,18 @@ public:
         if (saturation == 0)
             hue = 0;
         else {
-            double redc = ( (cmax - r)) / ((cmax - cmin));
-            double greenc = ( (cmax - g)) / ((cmax - cmin));
-            double bluec  = ((cmax - b)) / ((cmax - cmin));
+            float redc = ( (cmax - r)) / ((cmax - cmin));
+            float greenc = ( (cmax - g)) / ((cmax - cmin));
+            float bluec  = ((cmax - b)) / ((cmax - cmin));
             if (r == cmax)
                 hue = bluec - greenc;
             else if (g == cmax)
-                hue = 2.0 + redc - bluec;
+                hue = 2 + redc - bluec;
             else
-				hue = 4.0 + greenc - redc;
-            hue = hue / 6.0;
+				hue = 4 + greenc - redc;
+            hue = hue / 6;
             if (hue < 0)
-                hue = hue + 1.0;
+                hue = hue + 1;
         }
         
         int h = hue*255;
